Track subsurfaces damages
This prevents applications such as firefox from freezing when nothing
else is going on. Those applications use subsurfaces (for efficiency
reasons), which we were not keeping track of before.

Applications who use subsurfaces were working fine as long as other
applications were updating or the user was simply moving the mouse (and
thus damaging the whole screen anyway).

I'm not sure if listening to map and unmap events of subsurfaces is
necessary.

Most of this was inspired by sway.